requirement.txt
requests>=2.31.0
django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

